+++ uiScripts/ui_Main.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 ################################################################################
-## Form generated from reading UI file 'MainTZRkpV.ui'
+## Form generated from reading UI file 'MainEYihRx.ui'
 ##
 ## Created by: Qt User Interface Compiler version 6.1.0
 ##
@@ -78,71 +78,6 @@ class Ui_MainWindow(object):
         self.gridLayout_2 = QGridLayout(self.gridLayoutWidget_2)
         self.gridLayout_2.setObjectName(u"gridLayout_2")
         self.gridLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.label_5 = QLabel(self.gridLayoutWidget_2)
-        self.label_5.setObjectName(u"label_5")
-
-        self.gridLayout_2.addWidget(self.label_5, 6, 1, 1, 1)
-
-        self.label_2 = QLabel(self.gridLayoutWidget_2)
-        self.label_2.setObjectName(u"label_2")
-
-        self.gridLayout_2.addWidget(self.label_2, 1, 1, 1, 1)
-
-        self.lineEdit_3 = QLineEdit(self.gridLayoutWidget_2)
-        self.lineEdit_3.setObjectName(u"lineEdit_3")
-
-        self.gridLayout_2.addWidget(self.lineEdit_3, 2, 2, 1, 1)
-
-        self.pushButton_4 = QPushButton(self.gridLayoutWidget_2)
-        self.pushButton_4.setObjectName(u"pushButton_4")
-
-        self.gridLayout_2.addWidget(self.pushButton_4, 3, 4, 1, 1)
-
-        self.label_3 = QLabel(self.gridLayoutWidget_2)
-        self.label_3.setObjectName(u"label_3")
-
-        self.gridLayout_2.addWidget(self.label_3, 2, 1, 1, 1)
-
-        self.pushButton_2 = QPushButton(self.gridLayoutWidget_2)
-        self.pushButton_2.setObjectName(u"pushButton_2")
-
-        self.gridLayout_2.addWidget(self.pushButton_2, 1, 4, 1, 1)
-
-        self.horizontalLayout = QHBoxLayout()
-        self.horizontalLayout.setObjectName(u"horizontalLayout")
-        self.pushButton_7 = QPushButton(self.gridLayoutWidget_2)
-        self.pushButton_7.setObjectName(u"pushButton_7")
-
-        self.horizontalLayout.addWidget(self.pushButton_7)
-
-
-        self.gridLayout_2.addLayout(self.horizontalLayout, 7, 4, 1, 1)
-
-        self.lineEdit = QLineEdit(self.gridLayoutWidget_2)
-        self.lineEdit.setObjectName(u"lineEdit")
-
-        self.gridLayout_2.addWidget(self.lineEdit, 6, 2, 1, 1)
-
-        self.label_4 = QLabel(self.gridLayoutWidget_2)
-        self.label_4.setObjectName(u"label_4")
-
-        self.gridLayout_2.addWidget(self.label_4, 3, 1, 1, 1)
-
-        self.lineEdit_4 = QLineEdit(self.gridLayoutWidget_2)
-        self.lineEdit_4.setObjectName(u"lineEdit_4")
-
-        self.gridLayout_2.addWidget(self.lineEdit_4, 1, 2, 1, 1)
-
-        self.pushButton = QPushButton(self.gridLayoutWidget_2)
-        self.pushButton.setObjectName(u"pushButton")
-
-        self.gridLayout_2.addWidget(self.pushButton, 0, 4, 1, 1)
-
-        self.pushButton_3 = QPushButton(self.gridLayoutWidget_2)
-        self.pushButton_3.setObjectName(u"pushButton_3")
-
-        self.gridLayout_2.addWidget(self.pushButton_3, 2, 4, 1, 1)
-
         self.label = QLabel(self.gridLayoutWidget_2)
         self.label.setObjectName(u"label")
 
@@ -153,15 +88,70 @@ class Ui_MainWindow(object):
 
         self.gridLayout_2.addWidget(self.lineEdit_2, 3, 2, 1, 1)
 
-        self.lineEdit_5 = QLineEdit(self.gridLayoutWidget_2)
-        self.lineEdit_5.setObjectName(u"lineEdit_5")
+        self.label_4 = QLabel(self.gridLayoutWidget_2)
+        self.label_4.setObjectName(u"label_4")
 
-        self.gridLayout_2.addWidget(self.lineEdit_5, 0, 2, 1, 1)
+        self.gridLayout_2.addWidget(self.label_4, 3, 1, 1, 1)
+
+        self.lineEdit = QLineEdit(self.gridLayoutWidget_2)
+        self.lineEdit.setObjectName(u"lineEdit")
+
+        self.gridLayout_2.addWidget(self.lineEdit, 6, 2, 1, 1)
+
+        self.label_2 = QLabel(self.gridLayoutWidget_2)
+        self.label_2.setObjectName(u"label_2")
+
+        self.gridLayout_2.addWidget(self.label_2, 1, 1, 1, 1)
+
+        self.pushButton_4 = QPushButton(self.gridLayoutWidget_2)
+        self.pushButton_4.setObjectName(u"pushButton_4")
+
+        self.gridLayout_2.addWidget(self.pushButton_4, 3, 3, 1, 1)
+
+        self.pushButton_2 = QPushButton(self.gridLayoutWidget_2)
+        self.pushButton_2.setObjectName(u"pushButton_2")
+
+        self.gridLayout_2.addWidget(self.pushButton_2, 1, 3, 1, 1)
+
+        self.label_5 = QLabel(self.gridLayoutWidget_2)
+        self.label_5.setObjectName(u"label_5")
+
+        self.gridLayout_2.addWidget(self.label_5, 6, 1, 1, 1)
+
+        self.label_3 = QLabel(self.gridLayoutWidget_2)
+        self.label_3.setObjectName(u"label_3")
+
+        self.gridLayout_2.addWidget(self.label_3, 2, 1, 1, 1)
+
+        self.lineEdit_4 = QLineEdit(self.gridLayoutWidget_2)
+        self.lineEdit_4.setObjectName(u"lineEdit_4")
+
+        self.gridLayout_2.addWidget(self.lineEdit_4, 1, 2, 1, 1)
 
         self.pushButton_5 = QPushButton(self.gridLayoutWidget_2)
         self.pushButton_5.setObjectName(u"pushButton_5")
 
-        self.gridLayout_2.addWidget(self.pushButton_5, 6, 4, 1, 1)
+        self.gridLayout_2.addWidget(self.pushButton_5, 6, 3, 1, 1)
+
+        self.lineEdit_3 = QLineEdit(self.gridLayoutWidget_2)
+        self.lineEdit_3.setObjectName(u"lineEdit_3")
+
+        self.gridLayout_2.addWidget(self.lineEdit_3, 2, 2, 1, 1)
+
+        self.pushButton = QPushButton(self.gridLayoutWidget_2)
+        self.pushButton.setObjectName(u"pushButton")
+
+        self.gridLayout_2.addWidget(self.pushButton, 0, 3, 1, 1)
+
+        self.pushButton_3 = QPushButton(self.gridLayoutWidget_2)
+        self.pushButton_3.setObjectName(u"pushButton_3")
+
+        self.gridLayout_2.addWidget(self.pushButton_3, 2, 3, 1, 1)
+
+        self.lineEdit_5 = QLineEdit(self.gridLayoutWidget_2)
+        self.lineEdit_5.setObjectName(u"lineEdit_5")
+
+        self.gridLayout_2.addWidget(self.lineEdit_5, 0, 2, 1, 1)
 
         self.pushButton_6 = QPushButton(self.gridLayoutWidget_2)
         self.pushButton_6.setObjectName(u"pushButton_6")
@@ -227,17 +217,16 @@ class Ui_MainWindow(object):
         self.pushButton_8.setText(QCoreApplication.translate("MainWindow", u"New Virtual Machine", None))
         self.pushButton_9.setText(QCoreApplication.translate("MainWindow", u"Start Selected Virtual Machine", None))
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.tab), QCoreApplication.translate("MainWindow", u"Main", None))
-        self.label_5.setText(QCoreApplication.translate("MainWindow", u"qemu-system-mips64el path", None))
+        self.label.setText(QCoreApplication.translate("MainWindow", u"qemu-img Path", None))
+        self.label_4.setText(QCoreApplication.translate("MainWindow", u"qemu-system-ppc path", None))
         self.label_2.setText(QCoreApplication.translate("MainWindow", u"qemu-system-i386 path", None))
         self.pushButton_4.setText(QCoreApplication.translate("MainWindow", u"Browse", None))
-        self.label_3.setText(QCoreApplication.translate("MainWindow", u"qemu-system-x86_64 path", None))
         self.pushButton_2.setText(QCoreApplication.translate("MainWindow", u"Browse", None))
-        self.pushButton_7.setText(QCoreApplication.translate("MainWindow", u"Reset", None))
-        self.label_4.setText(QCoreApplication.translate("MainWindow", u"qemu-system-ppc path", None))
+        self.label_5.setText(QCoreApplication.translate("MainWindow", u"qemu-system-mips64el path", None))
+        self.label_3.setText(QCoreApplication.translate("MainWindow", u"qemu-system-x86_64 path", None))
+        self.pushButton_5.setText(QCoreApplication.translate("MainWindow", u"Browse", None))
         self.pushButton.setText(QCoreApplication.translate("MainWindow", u"Browse", None))
         self.pushButton_3.setText(QCoreApplication.translate("MainWindow", u"Browse", None))
-        self.label.setText(QCoreApplication.translate("MainWindow", u"qemu-img Path", None))
-        self.pushButton_5.setText(QCoreApplication.translate("MainWindow", u"Browse", None))
         self.pushButton_6.setText(QCoreApplication.translate("MainWindow", u"Apply", None))
         self.tabWidget_2.setTabText(self.tabWidget_2.indexOf(self.tab_3), QCoreApplication.translate("MainWindow", u"QEMU", None))
         self.label_6.setText(QCoreApplication.translate("MainWindow", u"EmuGUI", None))